+++ src/router/e2fsprogs/lib/ext2fs/tst_iscan.c
@@ -182,7 +182,8 @@ static void check_map(void)
 
 	for (i=0; test_vec[i]; i++) {
 		if (ext2fs_test_block_bitmap2(touched_map, test_vec[i])) {
-			printf("Bad block was touched --- %llu\n", test_vec[i]);
+			printf("Bad block was touched --- %llu\n",
+			       (unsigned long long) test_vec[i]);
 			failed++;
 			first_no_comma = 1;
 		}
@@ -194,7 +195,8 @@ static void check_map(void)
 		     j++, blk++) {
 			if (!ext2fs_test_block_bitmap2(touched_map, blk) &&
 			    !ext2fs_test_block_bitmap2(bad_block_map, blk)) {
-				printf("Missing block --- %llu\n", blk);
+				printf("Missing block --- %llu\n",
+				       (unsigned long long) blk);
 				failed++;
 			}
 		}
